+++ source/dem/dem.cc
@@ -181,7 +181,18 @@ DEMSolver<dim, PropertiesIndex>::setup_distribution_type()
               std::make_shared<NormalDistribution>(
                 lpp.particle_average_diameter.at(particle_type),
                 lpp.particle_size_std.at(particle_type),
-                lpp.seed_for_distributions[particle_type] + this_mpi_process);
+                lpp.seed_for_distributions[particle_type] + this_mpi_process,
+                lpp.diameter_min_cutoff[particle_type],
+                lpp.diameter_max_cutoff[particle_type]);
+            break;
+          case SizeDistributionType::lognormal:
+            size_distribution_object_container[particle_type] =
+              std::make_shared<LogNormalDistribution>(
+                lpp.particle_average_diameter.at(particle_type),
+                lpp.particle_size_std.at(particle_type),
+                lpp.seed_for_distributions[particle_type] + this_mpi_process,
+                lpp.diameter_min_cutoff[particle_type],
+                lpp.diameter_max_cutoff[particle_type]);
             break;
           case SizeDistributionType::custom:
             size_distribution_object_container[particle_type] =
@@ -191,6 +202,9 @@ DEMSolver<dim, PropertiesIndex>::setup_distribution_type()
                 lpp.seed_for_distributions[particle_type] + this_mpi_process);
             break;
         }
+      // To be uncommented in a follow-up PR since this will change all test
+      // outputs. size_distribution_object_container[particle_type]
+      //   ->print_psd_declaration_string(particle_type, pcout);
 
       maximum_particle_diameter = std::max(
         maximum_particle_diameter,